Add decimal support (#173)
* Use DecimalGraphType for decimals

* Version bump

* Remove version bump

* Refactor + assert result

+++ src/GraphQL.Conventions/Types/Resolution/TypeResolver.cs
@@ -78,7 +78,7 @@ namespace GraphQL.Conventions.Types.Resolution
             RegisterScalarType<ulong>(TypeNames.Integer);
             RegisterScalarType<float>(TypeNames.Float);
             RegisterScalarType<double>(TypeNames.Float);
-            RegisterScalarType<decimal>(TypeNames.Float);
+            RegisterScalarType<decimal>(TypeNames.Decimal);
             RegisterScalarType<DateTime>(TypeNames.DateTime);
             RegisterScalarType<DateTimeOffset>(TypeNames.DateTimeOffset);
             RegisterScalarType<TimeSpan>(TypeNames.TimeSpan);
